HOCS-5569: refactor unallocated accordion
Rework of the unallocated accordion to use the new endpoint within
casework. This also changes the functionality as we now derive
configuration per case type to enable and disable certain aspects
(displayAsFields etc.). This is effectively a merge of the old read-only
view and the all stages view.
