+++ server/config/unallocated/config.json
@@ -1,0 +1,7 @@
+{
+    "WCS": {
+      "displayAsFields": true,
+      "displayAll": true,
+      "displayHeading": false
+    }
+}
