+++ server/config/unallocated/unallocatedConfiguration.js
@@ -1,0 +1,12 @@
+const config = require('./config.json');
+
+const defaultConfig = {
+    displayAsFields: false,
+    displayAll: false,
+    displayHeading: true
+};
+
+const fetchUnallocatedConfigurationForCaseType = (caseType) =>
+    Object.assign({}, defaultConfig, config[caseType]);
+
+module.exports = fetchUnallocatedConfigurationForCaseType;
