--- server/lists/adapters/case-view-read-only.js
@@ -1,5 +1,0 @@
-const Form = require('../../services/forms/form-builder');
-
-module.exports = async (template) => {
-    return Form(template).withNoPrimaryAction().build();
-};
