+++ server/services/form.js
@@ -8,11 +8,9 @@ const FormBuilder = require('./forms/form-builder');
 const { Component } = require('./forms/component-builder');
 const { getSomuItem } = require('../middleware/somu');
 
-async function returnReadOnlyCaseViewForm(requestId, user, caseId) {
-    const { readOnlyCaseViewAdapter } = await listService
-        .getInstance(requestId, user)
-        .fetch('S_SYSTEM_CONFIGURATION');
-    return await listService.getInstance(requestId, user).fetch(readOnlyCaseViewAdapter, { caseId });
+async function returnUnallocatedCaseViewForm(requestId, user, caseId) {
+    return await listService.getInstance(requestId, user)
+        .fetch('CASE_VIEW_UNALLOCATED', { caseId });
 }
 
 async function getFormSchemaFromWorkflowService(requestId, options, user) {
@@ -26,7 +24,7 @@ async function getFormSchemaFromWorkflowService(requestId, options, user) {
             case 401:
                 // handle no permission to allocate
                 try {
-                    return { form: await returnReadOnlyCaseViewForm(requestId, user, caseId) };
+                    return { form: await returnUnallocatedCaseViewForm(requestId, user, caseId) };
                 } catch (error) {
                     return { error: new AuthenticationError('You are not authorised to work on this case') };
                 }
@@ -43,7 +41,7 @@ async function getFormSchemaFromWorkflowService(requestId, options, user) {
                     usersInTeam = [];
                 }
                 try {
-                    caseView = await returnReadOnlyCaseViewForm(requestId, user, caseId);
+                    caseView = await returnUnallocatedCaseViewForm(requestId, user, caseId);
                 } catch (error) {
                     caseView = null;
                 }
@@ -67,6 +65,7 @@ async function getFormSchemaFromWorkflowService(requestId, options, user) {
                     .withSecondaryAction(
                         Component('backlink')
                             .withProp('label', 'Cancel')
+                            .withProp('action', '/')
                             .build()
                     )
                     .build();
@@ -76,7 +75,7 @@ async function getFormSchemaFromWorkflowService(requestId, options, user) {
         }
     }
     if (response.data.form === null) {
-        const response = await returnReadOnlyCaseViewForm(requestId, user, caseId);
+        const response = await returnUnallocatedCaseViewForm(requestId, user, caseId);
         return { form: response };
     }
     const { stageUUID, caseReference, allocationNote } = response.data;
